build(dut): Gate lwIP DUT first-party TUs with strict warnings
- Demote lwIP src/contrib includes to -isystem on the apps
- Source-scope the strict set onto the 11 first-party C++ TUs
- Fix the ssize_t to int casts the gate surfaced

+++ dut/lwip_dut/lwip_reactor_test.cpp
@@ -202,7 +202,7 @@ int main() {
     sendTo(ctl, kControlPort, msg.data(), msg.size());
 
     std::uint8_t rbuf[256];
-    const int rn = lwip_recvfrom(ctl, rbuf, sizeof(rbuf), 0, nullptr, nullptr);
+    const ssize_t rn = lwip_recvfrom(ctl, rbuf, sizeof(rbuf), 0, nullptr, nullptr);
     check(rn >= static_cast<int>(tp::kHeaderSize), "control primitive got a response (waker path)");
     if (rn >= static_cast<int>(tp::kHeaderSize)) {
         const auto resp = tp::parseHeader(rbuf, static_cast<std::size_t>(rn));
